chore: organize app code
