+++ apps/app/src/components/login/KakaoLogin.tsx
@@ -6,7 +6,7 @@ import { login as kakaoLogin } from '@react-native-seoul/kakao-login'
 import { isErrorResponse } from '@fienmee/types'
 
 import { login } from '@/api'
-import { setToken } from '@/stores/token'
+import { setToken } from '@/stores'
 import { LoginScreenProps } from '@/types'
 
 export function KakaoOauthLogin() {
